refactor: split model file in differents files by struct type
